Refactor and reorganization of the existing interfaces
