+++ src/interfaces/Coordinate.ts
@@ -1,0 +1,4 @@
+export interface Coordinate {
+    lat: number;
+    lng: number;
+}
